Put function in different package to improve decopling.
