+++ src/test/java/drevo/stochastic/annealing/UnimodalFunctionsTest.java
@@ -6,52 +6,11 @@ import java.util.concurrent.ThreadLocalRandom;
 
 import org.junit.jupiter.api.Test;
 
+import drevo.stochastic.annealing.function.QuadraticFunction;
 import drevo.stochastic.annealing.monitoring.AnnealingListener;
 import drevo.stochastic.annealing.monitoring.AnnealingState;
 
 class UnimodalFunctionsTest extends BaseFunctionTest {
-    class QuadraticFunction implements AnnealingFunction {
-        ThreadLocalRandom rnd = ThreadLocalRandom.current();
-
-        private double x = rnd.nextDouble() * 20 - 10;
-    
-        public double getX() {
-            return x;
-        }
-
-        @Override
-        public double compute() {
-            return compute(x);
-        }
-        public double compute(double value) {
-            return (value - 2.0) * (value - 2.0);
-        }
-    
-        @Override
-        public void reconfigure() {
-            x = rnd.nextDouble() * 20 - 10; // Adjust x randomly within [-10, 10]
-        }
-    
-        @Override
-        public void assign(AnnealingFunction f) {
-            if (f instanceof QuadraticFunction) {
-                this.x = ((QuadraticFunction) f).x;
-            }
-        }
-    
-        @Override
-        public boolean isValid() {
-            return -10.0 <= x && x <= 10; // Valid domain: [-10, 10]
-        }
-    
-        @Override
-        public AnnealingFunction copy() {
-            QuadraticFunction clone = new QuadraticFunction();
-            clone.x = this.x;
-            return clone;
-        }
-    }
-
     @Test
     void minimizeQuadraticFunctionDefaultTest() {
         // Define a simple Function for testing
@@ -72,7 +31,7 @@ class UnimodalFunctionsTest extends BaseFunctionTest {
 
         // Assert that we found a reasonable solution
         assertTrue(Math.abs(expected - result.compute()) < 10e-5, String.format("It didn't minimize. expected: %.5f, result.compute(): %.5f.", expected, result.compute()));
-        assertTrue(Math.abs(expectedX - result.x) < 10e-2, String.format("The x value do not minimize (x - 2)^2. expectedX: %.5f, result.x: %.5f", expectedX, result.x));
+        assertTrue(Math.abs(expectedX - result.x()) < 10e-2, String.format("The x value do not minimize (x - 2)^2. expectedX: %.5f, result.x: %.5f", expectedX, result.x()));
     }
 
     @Test
@@ -95,6 +54,6 @@ class UnimodalFunctionsTest extends BaseFunctionTest {
 
         // Assert that we found a reasonable solution
         assertTrue(Math.abs(expected - result.compute()) < 10e-5, String.format("It didn't minimize. expected: %.5f, result.compute(): %.5f.", expected, result.compute()));
-        assertTrue(Math.abs(expectedX - result.x) < 10e-2, String.format("The x value do not minimize (x - 2)^2. expectedX: %.5f, result.x: %.5f", expectedX, result.x));
+        assertTrue(Math.abs(expectedX - result.x()) < 10e-2, String.format("The x value do not minimize (x - 2)^2. expectedX: %.5f, result.x: %.5f", expectedX, result.x()));
     }
 }
